refactor: move annotations to new package
